pyproject.toml
[tool.poetry]
name = "cs415-project-3"
version = "0.1.0"
description = ""
authors = ["Alan Bixby <alanrbixby@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
motor = "^3.1.1"
aiostream = "^0.4.5"
seaborn = "^0.12.1"
numpy = "^1.23.5"
matplotlib = "^3.6.2"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
thefuzz = {extras = ["speedup"], version = "^0.19.0"}
scipy = "^1.9.3"

[tool.poetry.group.lint.dependencies]
isort = "^5.10.1"
black = "^22.10.0"
flake8 = "^5.0.4"
mypy = "^0.982"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
notebook = "^6.5.2"
ipykernel = "^6.17.1"

[tool.isort]
profile = "black"

[tool.mypy]
# 3rd party import
ignore_missing_imports = true
# dynamic typing
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = false
disallow_any_generics = false
disallow_subclassing_any = true
# platform
python_version = "3.11"
# untyped
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
# None and Optional
no_implicit_optional = true
# Warnings
warn_return_any = false
warn_unreachable = true
# Misc
pretty = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
